Remove dependency on cpptest.
As the variation on testing required is very small
(two simple macros is enough), the unittesting library
was removed by a local implementation.
This makes it easier for anyone who compiled the library
to make sure it works properly against the included test suite.

Clean 'doc' and 'tests' aswell when issuing clean.