Implementation of serialize and unserialize.
To provide an easy to use interface, the basic functions
serialize and deserialize is introduced.

Deserialize is an exact copy of Json::Builder::parse, therefore
the parse function was removed.

Serialize takes "any" object and returns its json representation.
This is not an exact copy of Json::Builder::create (where Json::Value
was returned). create is kept intact.